+++ packages/showcases/web-components/.storybook/main.js
@@ -1,6 +1,13 @@
 module.exports = {
   stories: ['../stories/**/*.stories.@(js|jsx|ts|tsx)'],
-  addons: ['@storybook/addon-essentials'],
+  addons: [
+    '@storybook/addon-essentials',
+    '@storybook/addon-a11y',
+    '@whitespace/storybook-addon-html',
+    'storybook-addon-designs',
+    'storybook-mobile',
+    'storybook-addon-outline',
+  ],
   webpack: (config) => {
     config.module.rules.push({
       test: /\.js$/,