+++ packages/showcases/web-components/.storybook/preview.js
@@ -1,8 +1,17 @@
 import '@vtmn/css';
+import { withDesign } from 'storybook-addon-designs';
+import backgrounds from '@vtmn/showcase-core/addons/backgrounds.json';
+import viewports from '@vtmn/showcase-core/addons/viewports.json';
 import { defineCustomElements } from '@vtmn/web-components/dist/esm/loader';
 
 defineCustomElements();
 
+export const decorators = [withDesign];
+
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
+  backgrounds,
+  viewport: {
+    viewports,
+  },
 };