+++ packages/sources/web-components/src/vtmn-button/vtmn-button.tsx
@@ -6,6 +6,10 @@ import { Component, Prop, h } from '@stencil/core';
   shadow: true,
 })
 export class VtmnButton {
+  /**
+   * The variant of the button.
+   * @default 'primary'
+   */
   @Prop() variant:
     | 'primary'
     | 'primary-reversed'
@@ -14,10 +18,16 @@ export class VtmnButton {
     | 'ghost'
     | 'conversion' = 'primary';
 
+  /**
+   * The size of the button.
+   * @default 'medium'
+   */
+  @Prop() size: 'small' | 'medium' | 'large' | 'stretched' = 'medium';
+
   render() {
     return (
       <button
-        class={`vtmn-btn vtmn-btn_variant--${this.variant}`}
+        class={`vtmn-btn vtmn-btn_variant--${this.variant} vtmn-btn_size--${this.size}`}
         type="button">
         <slot></slot>
       </button>