+++ packages/sources/web-components/stencil.config.ts
@@ -16,10 +16,6 @@ export const config: Config = {
     {
       type: 'docs-readme',
     },
-    {
-      type: 'www',
-      serviceWorker: null, // disable service workers
-    },
   ],
   plugins: [
     postcss({